Fixes: - definitions for windows - subscription_events unit test (use sync client iterate) - corpus generation - access after free - #2394 - fuzzing cmake - pubsub includes - double free - amalgamation - uninitialized warning on windows - docs - fuzzing tests - extern declarations - includes - casting error - undefined behavior in historydatabse (memcpy called with null) - client and client_async after rebasing. - check_client - check_client_async - check_client_subscriptions - check_accesscontrol - check_client_async_connect - check_client_securechannel - check_services_view - check_monitoreditem_filter - check_subscription_events - check_server_historical_data - Fix check_discovery - check_pubsub_informationmodel_methods - check_client_highlevel - check_client_historical_data - encryption tests - includes - #2416

+++ plugins/networking/ua_select_based_networkmanager.c
@@ -2,11 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  *
- *    Copyright 2018 (c) Mark Giraud, Fraunhofer IOSB
+ *    Copyright 2018-2019 (c) Mark Giraud, Fraunhofer IOSB
  */
 
 #include "open62541_queue.h"
 #include "ua_plugin_log.h"
+#include "ua_plugin_socket.h"
+#include "ua_plugin_network_manager.h"
 #include "ua_types_generated_handling.h"
 #include "ua_networkmanagers.h"
 
@@ -144,7 +146,7 @@ select_nm_processSocket(UA_NetworkManager *networkManager, UA_UInt32 timeout,
     fd_set fdset;
     FD_ZERO(&fdset);
     UA_fd_set((UA_SOCKET)sock->id, &fdset);
-    struct timeval tmptv = {0, timeout * 1000};
+    struct timeval tmptv = {0, (long int)(timeout * 1000)};
 
     int resultsize = UA_select((UA_Int32)(sock->id + 1), NULL, &fdset, NULL, &tmptv);
     UA_StatusCode retval = UA_STATUSCODE_GOOD;
